[klibc] ppc64: fix struct stat
On ppc64 the struct stat defined by klibc matches the kernel one.
However it contains implicit padding before the st_rdev field due to the
64-bit alignement. For internal reasons, klibc defines st_rdev as a pair
of 32-bit values (using the __stdev64 macro). They only need to be
32-bit aligned and as a consequence st->st_rdev is incorrectly defined.

The solution is to add an explicit padding in the structure. This fixes
the resume binary on ppc64 BE and LE, and probably other things.

Signed-off-by: Aurelien Jarno <aurelien@aurel32.net>
Link: http://lkml.kernel.org/r/20160106004416.GG28542@decadent.org.uk
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>
Signed-off-by: H. Peter Anvin <hpa@linux.intel.com>