[klibc] run-init: Add dry-run mode
initramfs-tools wants to validate the real init program before running
it, as there is no way out once it has exec'd run-init.  This is
complicated by the increasing use of symlinks for /sbin/init and for
/sbin itself.  We can't simply resolve them with 'readlink -f' because
any absolute symlinks will be resolved using the wrong root.  Add a
dry-run mode (-n option) to run-init that goes as far as possible to
validate that the given init is executable.

References: https://bugs.debian.org/810965
Link: https://www.zytor.com/pipermail/klibc/2017-December/003973.html
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>