[klibc] ipconfig: handle multiple interfaces correctly
	 When configuring multiple interfaces, the existing logic in
ipconfig can fail if DHCP replies are received out of the expected order,
or if one or more interfaces never receive replies.

	The current ipconfig logic uses a single packet socket to handle
all incoming DHCP replies.  If, for example, the host has two interfaces,
A and B, and only B will be sent a DHCP reply, the order of events goes
something like this:

ipconfig sends DHCP on A
ipconfig sends DHCP on B
system receives DHCP response to B
ipconfig attempts to receive DHCP reply for A; the available packet does
not match A's XID, et al, and is discarded.

	This will repeat until ipconfig times out and gives up.

	The failure can come and go with different kernel versions, as the
kernel version affects the order in which the devices appear.  In the
above example, if interface B is first, then the sequence will work.

	The solution is to use a packet socket per interface.  This avoids
the need to queue packets being read for the "wrong" interface or other
schemes to peek incoming packets.

Signed-off-by: Jay Vosburgh <jay.vosburgh@canonical.com>
Link: https://www.zytor.com/pipermail/klibc/2017-February/003945.html
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>