[klibc] Add support for reboot syscall argument
Add support to the reboot command so an argument for the 4th parameter
of the syscall can be passed around. This can be useful in, say, devices
that follow Android partitions conventions, so we can do things like
rebootinig to the kernel in recovery partition with "reboot recovery".

References: https://bugs.launchpad.net/ubuntu/+source/klibc/+bug/1692494
Link: https://www.zytor.com/pipermail/klibc/2017-May/003957.html
Link: https://bugs.debian.org/863761
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>