[klibc] Fix sparc assembly when compiled as PIC
Some distributions default to PIE for their compilers, which on sparc is passed
on to the assembler. Since the behaviour of %hi/%lo changes under PIC to become
GOT offsets, the current assembly files need adapting to not try to use a GOT
offset as an absolute address.

References: https://bugs.debian.org/885852
Link: https://www.zytor.com/pipermail/klibc/2018-July/004001.html
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>