[klibc] Build and install kinit and sh without ".shared" suffix
Most binaries are installed with the same name regardless of
whether they use a static or shared library.  The two exceptions
to this are kinit and sh, which are installed with the suffix
".shared" if they use a shared library.

Build these binaries in subdirectories, so that the static and shared
versions have the same (unsuffixed) name.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>