[klibc] Fix errlist.c generation in out-of-tree build
Kbuild expands every -I in the compiler options so that in an
out-of-tree build both the source directory and corresponding
object directory are on the include path.

We currently pass $(KLIBCCPPFLAGS) to makeerrlist.pl, which is not
expanded, so in an out-of-tree buld it only looks in the object
directories and does not find linux/errno.h.  Use the flags macro to
expand $(KLIBCCPPFLAGS).

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>