[klibc] alpha: Fix division routine generation in out-of-tree build
Kbuild expands every -I in the compiler options so that in an
out-of-tree build both the source directory and corresponding
object directory are on the include path.

We currently use $(KLIBCCPPFLAGS) when compiling divide.c, which is
not expanded, so in an out-of-tree buld the compiler does not find
headers in the source tree.  Use the flags macro to expand
$(KLIBCCPPFLAGS).

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>