[klibc] run-init: Allow the initramfs to be persisted across root changes
systemd supports switching back to the initramfs during shutdown in
order to make it easier to clean up the root file system. This is
desirable in order to allow us to remove keys from RAM before rebooting,
making it harder to obtain confidential information by rebooting into an
environment that scrapes RAM contents.

Signed-off-by: Matthew Garrett <mjg59@google.com>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>