[klibc] losetup: Use LOOP_CTL_GET_FREE to find free device
Since Linux 3.1, the loop driver creates a /dev/loop-control device
node which supports ioctls to allocate and free devices.

When the loop driver is modular, udev creates this in advance, and
opening it causes the driver to be loaded.  (The same is not true for
/dev/loop*.)

Using the LOOP_CTL_GET_FREE ioctl also allows creating more than the
default number of loop devices, and is more efficient than checking
a range of possible device names.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>