[klibc] Kbuild: Tell gas we don't want executable stacks
The stack should be made non-executable, as a security hardening
measure.  This is irrelevant for most of the klibc utilities, but
ipconfig deals with network input that might be untrusted.  Since
Linux 5.6-rc1 the kernel now also warns (once) if a program has an
executable stack.

As this is necessarily a process-wide attribute at run-time, the
stack ends up being executable unless every object file linked into the
program is flagged as not needing it.  gas doesn't set the flag by
default, so we need to explicitly tell it to do so.

ia64 will also need a change to its linker script to retain the
.note.GNU-stack section, but I have no way of testing ia64 so I'm going
to leave that to later.

Reported-by: Christophe Leroy <christophe.leroy@c-s.fr>
References: https://lists.zytor.com/archives/klibc/2020-February/004271.html
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>