[klibc] ipconfig: Fix alignment of packet structure
This structure should have alignment of exactly 4, since we want the
16-bit and 32-bit fields in it to have natural alignment but we don't
want extra padding that would violate the wire protocols.

This should be what happens by default, but for some reason it's
declared with the "packed" attribute.  gcc is now warning about using
potentially misaligned pointers to its members.

Add the attribute "aligned(4)" so that it is definitely exactly 4-byte
aligned.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>