klcc: Remove the need for bash
Currently, in the entire klibc bash is only used to identify the path of
the perl binary. It is doing so using the bash built-in 'type' function,
which is POSIX compliant according to [0], but the option -c is not.

By using `command -v` instead, we achieve the same result, in a POSIX
compliant manor [1], potentially removing the unneeded bash dependency.

0 https://www.unix.com/man-page/posix/1p/type/
1 https://www.unix.com/man-page/posix/1p/command/

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>