[klibc] stdio: Define all the _unlocked functions and macros
Clang 9.0 can optimise stdio function calls to use the _unlocked
variants, presumably based on escape analysis.  We don't define many
of them, and adding a lot of -fno-builtin-* options to inhibit this
seems like a losing battle.

Since we don't support multi-threaded programs or locking in stdio,
define the _unlocked function names as aliases.

For completeness, also define corresponding _unlocked macros for the
macro-only pseudo-functions, and for the inline functions in stdio.h.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>