[klibc] riscv64: Make linker relaxation work and enable it
RISC-V code generally needs two instructions to address static data,
but up to 4K of static data in an executable can be addressed using a
single instruction.  The linker "relaxes" a two-instruction sequence
to one instruction, assuming that the gp register is initialised
appropriately.

In my initial port, I had this working for static executables but not
for those using the shared library, so I disabled relaxation.
gp-relative relaxation is *not* meant to be used in shared libraries,
but we should enable it in executables.  pc-relative relaxation can be
enabled in all cases.

* When linking the shared library, define __global_pointer$ as 0 to
  disable gp-relative relaxation
* For executables using the shared-library, add a _main routine in
  _main.S that initialises gp and tail-calls main
* Add _main.o to $(KLIBCCRTSHARED) and change the entry point to _main
* Install _main.o so klcc can use it
* Drop the --no-relax linker option

Thanks: Jessica Clarke <jrtc27@jrtc27.com>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>