[klibc] klcc: Force make to use a shell to run 'command'
GNU make has an optimisation to run some commands without using the
shell, and its rules for recognising whether the shell is needed can
be buggy.

In particular, older versions don't recognise that 'command' may be a
shell built-in, and try to run it as if it's an external command.  Add
a redirection to the commamd line, to inhibit this optimisation.

Reported-by: Greg Thelen <gthelen@google.com>
References: https://lists.zytor.com/archives/klibc/2020-September/004548.html
References: https://git.kernel.org/linus/913ab9780fc021298949cc5514d6255a008e69f9
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>